move new tasks and start to add tests
